requirements.txt
Django>=5.1
djangorestframework>=3.15
psycopg2-binary>=2.9
